added utilities
