+++ src/idvc/dvc_interface.py
@@ -21,12 +21,12 @@ from PySide2.QtCore import (QByteArray, QRegExp, QSettings, QSize, Qt,
                             QThreadPool)
 from PySide2.QtGui import QCloseEvent, QKeySequence, QRegExpValidator
 from PySide2.QtWidgets import (QAction, QCheckBox, QComboBox,
-                               QDialog, QDialogButtonBox, QDockWidget,
+                               QDockWidget,
                                QDoubleSpinBox, QFileDialog, QFormLayout,
                                QFrame, QGroupBox, QLabel, QLineEdit,
                                QMainWindow, QMessageBox,
                                QProgressDialog, QPushButton, QSpinBox,
-                               QStatusBar, QStyle, QTabWidget, QVBoxLayout,
+                               QTabWidget, QVBoxLayout,
                                QHBoxLayout, QSizePolicy,
                                QWidget)
 import time
@@ -90,60 +90,13 @@ from idvc import version as gui_version
 
 from idvc.ui.Widgets import *
 from idvc.ui.Dialogs import * 
+from idvc.utilities import *
 
 __version__ = gui_version.version
 
 import logging
 
-class PrintCallback(object):
-    '''Class to handle the emit call when no callback is provided'''
-    def emit(self, *args, **kwargs):
-        print (args, kwargs)
 
-def reduce_displ(raw_displ, min_size, max_size, pzero=False):
-    '''filter the diplacement vectors based on their size'''
-    offset = 6 # 6 in the case of the iDVC
-    
-    # sizes = []
-    # dmin = np.inf
-    # dmax = 0.
-    # for el in raw_displ:
-    #     size = 0
-    #     for i in range(3):
-    #         #calculate size of vector
-    #         size += el[i+offset]*el[i+offset]
-    #     size = np.sqrt(size)
-    #     if size > dmax:
-    #         dmax = size
-    #     if size < dmin:
-    #         dmin = size
-    #     sizes.append(size)
-    vec = np.asarray(raw_displ)[:,offset:offset+3]
-    if pzero:
-        vec -= np.asarray(raw_displ[0][offset:offset+3])
-
-    sizes = np.sqrt( np.sum( np.power(vec, 2), axis=1) )
-
-    dmin = sizes.min()
-    dmax = sizes.max()
-    if min_size is None and max_size is None:
-        displ = raw_displ
-    else:
-        displ = []
-        if pzero:
-            for i in range(len(raw_displ)):
-                size = sizes[i]
-                if size > min_size  and size < max_size :
-                    line = raw_displ[i]
-                    line[offset:offset+3] -= vec[0]
-                    displ.append(line)
-        else:
-            for i in range(len(raw_displ)):
-                size = sizes[i]
-                if size > min_size  and size < max_size :
-                    displ.append(raw_displ[i])
-    displ = np.asarray(displ)
-    return displ, dmin, dmax
 
 class MainWindow(QMainWindow):
     def __init__(self):
@@ -5487,450 +5440,3 @@ Please select the new location of the file, or move it back to where it was orig
 
 
         
-
-
-class SaveObjectWindow(QtWidgets.QWidget):
-    '''a window which will appear when saving a mask or pointcloud
-    '''
-        #self.copy_files_label = QLabel("Allow a copy of the image files to be stored: ")
-
-    def __init__(self, parent, object_type, save_only):
-        super().__init__()
-
-        #print(save_only)
-
-        self.parent = parent
-        self.object = object_type
-
-        if self.object == "mask":
-            self.setWindowTitle("Save Existing Mask")
-            self.label = QLabel("Save mask as:")
-        elif self.object == "pointcloud":
-            self.setWindowTitle("Save Existing Point Cloud")
-            self.label = QLabel("Save Point Cloud as:")
-
-
-        self.setWindowModality(QtCore.Qt.ApplicationModal)
-        #self.setInputMode(QtWidgets.QInputDialog.TextInput)
-
-        self.textbox = QLineEdit(self)
-        rx = QRegExp("[A-Za-z0-9]+")
-        validator = QRegExpValidator(rx, self.textbox) #need to check this
-        self.textbox.setValidator(validator)
-
-        self.save_button = QPushButton("Save")
-        self.quit_button = QPushButton("Discard")
-        self.save_button.clicked.connect(lambda: self.save(save_only))
-        self.quit_button.clicked.connect(self.quit)
-        
-        self.setWindowFlags(QtCore.Qt.WindowTitleHint )
-        #self.setCancelButtonText("New Session")
-        #self.setAttribute(Qt.WA_DeleteOnClose)
-        self.layout = QtWidgets.QFormLayout()
-        self.layout.addRow(self.label)
-        self.layout.addRow(self.textbox)
-        self.layout.addRow(self.save_button, self.quit_button)
-        self.setLayout(self.layout)
-
-
-    def save(self, save_only):
-        if self.object == "mask":
-            #Load Saved Session
-            #print("Write mask to file, then carry on")
-            filename = self.textbox.text() + ".mha"
-            shutil.copyfile(os.path.join(tempfile.tempdir, self.parent.mask_file), os.path.join(tempfile.tempdir, "Masks", filename))
-            self.parent.mask_parameters['masksList'].addItem(filename)
-            self.parent.mask_details[filename] = self.parent.mask_details['current']
-            #print(self.parent.mask_details)
-
-            self.parent.mask_parameters['loadButton'].setEnabled(True)
-            self.parent.mask_parameters['masksList'].setEnabled(True)
-
-
-            if not save_only:
-                #print("Not save only")
-                #would be better to move this elsewhere
-                self.parent.mask_worker = Worker(self.parent.extendMask)
-                self.parent.create_progress_window("Loading", "Loading Mask")
-                self.parent.mask_worker.signals.progress.connect(self.parent.progress)
-                self.parent.mask_worker.signals.finished.connect(self.parent.DisplayMask)
-                self.parent.threadpool.start(self.parent.mask_worker)
-                self.parent.progress_window.setValue(10)
-            
-        if self.object == "pointcloud":
-            filename = self.textbox.text() + ".roi"
-            shutil.copyfile(os.path.join(tempfile.tempdir, "latest_pointcloud.roi"), os.path.join(tempfile.tempdir, filename))
-
-            self.parent.pointcloud_parameters['loadButton'].setEnabled(True)
-            self.parent.pointcloud_parameters['pointcloudList'].setEnabled(True)
-            self.parent.pointcloud_parameters['pointcloudList'].addItem(filename)
-            self.parent.pointCloud_details[filename] = self.parent.pointCloud_details['latest_pointcloud.roi']
-            #print(self.parent.pointCloud_details)
-            #self.parent.createPointCloud()
-            if not save_only:
-                self.parent.PointCloudWorker("create")
-            
-
-        self.close()
-
-    def quit(self):
-        if self.object == "mask":
-            #would be better to move this elsewhere
-            self.parent.mask_worker = Worker(self.parent.extendMask)
-            self.parent.create_progress_window("Loading", "Loading Mask")
-            self.parent.mask_worker.signals.progress.connect(self.parent.progress)
-            self.parent.mask_worker.signals.finished.connect(self.parent.DisplayMask)
-            self.parent.threadpool.start(self.parent.mask_worker)
-            self.parent.progress_window.setValue(10)
-
-        if self.object == "pointcloud":
-            self.parent.PointCloudWorker("create")
-            #self.parent.createPointCloud()
-
-        self.close()
-
-class VisualisationWindow(QtWidgets.QMainWindow):
-    '''creates a window which will contain the VisualisationWidgets
-    '''
-    def __init__(self, parent):
-        super().__init__()
-        self.parent = parent
-        self.setMinimumSize(200,200)
-
-class VisualisationWidget(QtWidgets.QMainWindow):
-    '''creates a window with a QCILViewerWidget as the central widget
-    '''
-    def __init__(self, parent, viewer=viewer2D, interactorStyle=vlink.Linked2DInteractorStyle):
-        super().__init__()
-        self.parent = parent
-
-        self.e = ErrorObserver()
-        self.viewer = viewer
-        self.interactorStyle = interactorStyle
-        self.createEmptyFrame()
-        self.threadpool = QThreadPool()
-
-    def getViewer(self):
-        return self.frame.viewer
-
-    def getInteractor(self):
-        return self.getViewer().getInteractor()
-
-    def getInteractorStyle(self):
-        return self.getViewer().style
-
-    def getViewerType(self):
-        return self.viewer
-
-        
-    def createEmptyFrame(self):
-        #print("empty")
-        self.frame = QCILViewerWidget(viewer=self.viewer, shape=(600,600), interactorStyle=self.interactorStyle)
-        self.setCentralWidget(self.frame)
-        self.image_file = [""]
-       
-    def displayImageData(self):
-        self.createEmptyFrame()
-        if self.viewer == viewer3D:
-            #set volume mapper according to user settings:
-            if self.parent.settings.value("volume_mapper") == "cpu":
-                self.frame.viewer.volume_mapper = vtk.vtkFixedPointVolumeRayCastMapper()
-                self.frame.viewer.volume.SetMapper(self.frame.viewer.volume_mapper)
-        else:
-            self.frame.viewer.setVisualisationDownsampling(self.parent.resample_rate)
-            self.frame.viewer.setDisplayUnsampledCoordinates(True)
-
-            vs_widgets = self.parent.visualisation_setting_widgets
-
-            vs_widgets['loaded_image_dims_value'].setVisible(True)
-            vs_widgets['loaded_image_dims_value'].setText(str(self.parent.unsampled_image_dimensions))
-
-            #print("resample rate: ", self.parent.resample_rate)
-
-            if self.parent.resample_rate != [1,1,1]:
-                vs_widgets['displayed_image_dims_value'].setVisible(True)
-                vs_widgets['displayed_image_dims_label'].setVisible(True)
-                #print("Disp image size ", [self.parent.ref_image_data.GetDimensions()[i] for i in range(3)])
-                vs_widgets['displayed_image_dims_value'].setText(str([round(self.parent.ref_image_data.GetDimensions()[i]) for i in range(3)]))
-                vs_widgets['coords_combobox'].setEnabled(True)
-                vs_widgets['coords_combobox'].setCurrentIndex(0)
-                vs_widgets['coords_warning_label'].setVisible(True)
-                vs_widgets['coords_info_label'].setVisible(True)
-
-            
-            else:
-                vs_widgets['displayed_image_dims_value'].setVisible(False)
-                vs_widgets['displayed_image_dims_label'].setVisible(False)
-                vs_widgets['coords_warning_label'].setVisible(False)
-                vs_widgets['coords_info_label'].setVisible(False)
-
-                vs_widgets['coords_combobox'].setEnabled(False)
-                vs_widgets['coords_combobox'].setCurrentIndex(0)
-
-        self.frame.viewer.setInput3DData(self.image_data)  
-        interactor = self.frame.viewer.getInteractor()
-
-
-        if hasattr(self.parent, 'orientation'):
-                orientation = self.parent.orientation
-        else:
-            orientation = self.frame.viewer.getSliceOrientation()
-        
-        if orientation == SLICE_ORIENTATION_XZ:
-            axis = 'y'
-        elif orientation == SLICE_ORIENTATION_YZ:
-            axis = 'x'
-        else:
-            axis = 'z'
-        interactor.SetKeyCode(axis)
-
-        if self.viewer == viewer2D:
-            self.frame.viewer.style.OnKeyPress(interactor, 'KeyPressEvent')
-            if self.parent.current_slice:
-                if self.parent.current_slice <= self.frame.viewer.img3D.GetExtent()[self.frame.viewer.getSliceOrientation()*2+1]:
-                    self.frame.viewer.displaySlice(self.parent.current_slice)
-
-
-        if self.viewer == viewer3D:
-            self.frame.viewer.style.OnKeyPress(interactor, 'KeyPressEvent')
-            # Depth peeling for volumes doesn't work as we would like when we have the vtk.vtkFixedPointVolumeRayCastMapper() instead of the vtk.vtkSmartVolumeMapper()
-            # self.frame.viewer.sliceActor.GetProperty().SetOpacity(0.99)
-            # self.frame.viewer.ren.SetUseDepthPeeling(True)
-            # self.frame.viewer.renWin.SetAlphaBitPlanes(True)
-            # self.frame.viewer.renWin.SetMultiSamples(False)
-            # self.frame.viewer.ren.UseDepthPeelingForVolumesOn()
-    
-            if self.parent.current_slice:
-                if self.parent.current_slice <= self.frame.viewer.img3D.GetExtent()[self.frame.viewer.getSliceOrientation()*2+1]:
-                    self.frame.viewer.style.SetActiveSlice(self.parent.current_slice)
-                    self.frame.viewer.style.UpdatePipeline()
-
-        # print("set input data for" + str(self.viewer))
-
-        if self.viewer == viewer2D:
-            self.PlaneClipper = cilPlaneClipper()
-            self.PlaneClipper.SetInteractorStyle(self.frame.viewer.style)
-
-
-    def setImageData(self, image_data):
-        self.image_data = image_data
-
-    def getImageData(self):
-        return self.image_data
-
-class GraphsWindow(QMainWindow):
-    '''creates a new window with graphs from results saved in the selected run folder.
-    '''
-    def __init__(self, parent=None):
-        super(GraphsWindow, self).__init__(parent)
-        self.setWindowTitle("Digital Volume Correlation Results")
-        DVCIcon = QtGui.QIcon()
-        DVCIcon.addFile("DVCIconSquare.png")
-
-        # Menu
-        self.menu = self.menuBar()
-        self.file_menu = self.menu.addMenu("File")
-        self.settings_menu = self.menu.addMenu("Settings")
-
-        displacement_setting_action = QAction("Show Displacement Relative to Reference Point 0", self)
-        displacement_setting_action.setCheckable(True)
-        displacement_setting_action.setChecked(False)
-        self.displacement_setting_action = displacement_setting_action
-
-        displacement_setting_action.triggered.connect(self.ReloadGraphs)
-        self.settings_menu.addAction(displacement_setting_action)
-
-
-        # Exit QAction
-        exit_action = QAction("Exit", self)
-        exit_action.setShortcut(QKeySequence.Quit)
-        exit_action.triggered.connect(self.close)
-        self.file_menu.addAction(exit_action)
-
-        #Tab positions:
-        self.setTabPosition(QtCore.Qt.AllDockWidgetAreas,QTabWidget.North)
-        self.setDockOptions(QMainWindow.ForceTabbedDocks)
-             
-        # Window dimensions
-        geometry = qApp.desktop().availableGeometry(self)
-
-        self.setGeometry(50,50, geometry.width()-100, geometry.height()-100)
-        #self.setFixedSize(geometry.width() * 0.6, geometry.height() * 0.8)
-
-    def SetResultsFolder(self, folder):
-        self.results_folder = folder
-        self.setWindowTitle("Digital Volume Correlation Results - {foldername}".format(foldername=os.path.basename(self.results_folder)))
-    
-    def ReloadGraphs(self):
-        self.DeleteAllWidgets()
-        self.CreateDockWidgets(displ_wrt_point0 = self.displacement_setting_action.isChecked())
-
-    def DeleteAllWidgets(self):
-         for current_dock in self.findChildren(QDockWidget):
-            current_dock.close()
-            del current_dock
-
-    def CreateDockWidgets(self, displ_wrt_point0 = False):
-        result_list=[]
-        #print(results_folder[0])
-        for folder in glob.glob(os.path.join(self.results_folder, "dvc_result_*")):
-            file_path = os.path.join(folder, os.path.basename(folder))
-            result = RunResults(file_path)
-            result_list.append(result)
-    
-            GraphWidget = SingleRunResultsWidget(self, result, displ_wrt_point0)
-            dock1 = QDockWidget(result.title,self)
-            dock1.setAllowedAreas(QtCore.Qt.RightDockWidgetArea)
-            dock1.setWidget(GraphWidget)
-            self.addDockWidget(QtCore.Qt.RightDockWidgetArea,dock1)
-    
-        prev = None
-
-        for current_dock in self.findChildren(QDockWidget):
-            if self.dockWidgetArea(current_dock) == QtCore.Qt.RightDockWidgetArea:
-                existing_widget = current_dock
-
-                if prev:
-                    self.tabifyDockWidget(prev,current_dock)
-                prev= current_dock
-        
-        SummaryTab = SummaryGraphsWidget(self, result_list)
-        dock = QDockWidget("Summary",self)
-        dock.setAllowedAreas(QtCore.Qt.RightDockWidgetArea)
-        dock.setWidget(SummaryTab)
-        self.addDockWidget(QtCore.Qt.RightDockWidgetArea,dock)
-        self.tabifyDockWidget(prev,dock)
-
-        dock.raise_() # makes summary panel the one that is open by default.
-
-
-        
-class RunResults(object):
-    def __init__(self, file_name):
-        
-        self.points = None
-
-        disp_file_name = file_name + ".disp"
-        stat_file_name = file_name + ".stat"
-
-        with open(stat_file_name,"r") as stat_file:
-            
-            count = 0
-            offset = 0
-            for line in stat_file:
-                if count == 9:
-                    if line.split('\t')[0] == "vol_endian":
-                        offset = 1
-
-                if count == 14 + offset:
-                    self.subvol_geom = str(line.split('\t')[1])
-                if count == 15 + offset:
-                    self.subvol_size = round(int(line.split('\t')[1]))
-                if count == 16 +offset:
-                    self.subvol_points = int(line.split('\t')[1])
-                if count == 20 + offset:
-                    self.disp_max = int(line.split('\t')[1])
-                if count == 21 + offset:
-                    self.num_srch_dof = int(line.split('\t')[1])
-                if count == 22 + offset:
-                    self.obj_function = str(line.split('\t')[1])
-                if count == 23 + offset:
-                    self.interp_type = str(line.split('\t')[1])
-                if count == 25 + offset:
-                    self.rigid_trans = [int(line.split('\t')[1]),int(line.split('\t')[2]), int(line.split('\t')[3])]
-                # if count == 26 + offset:
-                #     self.basin_radius = int(line.split('\t')[1])
-                # if count == 27 + offset:
-                #     self.subvol_aspect = [int(line.split('\t')[1]),int(line.split('\t')[2]), int(line.split('\t')[3])]
-                count+=1
-
-        plot_titles_dict = {
-            'objmin': "Objective Minimum", 'u': "Displacement in x", 'v':"Displacement in y", 'w':"Displacement in z",
-            'phi':"Change in phi",'the':"Change in theta", 'psi':"Change in psi"}
-
-        with open(disp_file_name) as f:
-            # first 4 columns are: n, x, y, z, status - we don't want these
-            self.plot_titles = f.readline().split()[5:]
-            self.plot_titles = [plot_titles_dict.get(text, text) for text in self.plot_titles]
-
-        
-        self.disp_file = disp_file_name
-
-        self.title =  str(self.subvol_points) + " Points in Subvolume," + " Subvolume Size: " + str(self.subvol_size)
-
-    def __str__(self):
-
-        a = "subvol_size {}".format(self.subvol_size)
-        n = "subvol_points {}".format(self.subvol_points)
-        return "RunResults:\n{}\n{}".format(a , n)
-
-def generateUIDockParameters(self, title): #copied from dvc_configurator.py
-    '''creates a dockable widget with a form layout group to add things to
-
-    basically you can add widget to the returned groupBoxFormLayout and paramsGroupBox
-    The returned dockWidget must be added with
-    self.addDockWidget(QtCore.Qt.RightDockWidgetArea, dockWidget)
-    '''
-    dockWidget = QDockWidget(self)
-    dockWidget.setWindowTitle(title)
-    dockWidgetContents = QWidget()
-
-
-    # Add vertical layout to dock contents
-    dockContentsVerticalLayout = QVBoxLayout(dockWidgetContents)
-    dockContentsVerticalLayout.setContentsMargins(0, 0, 0, 0)
-
-    # Create widget for dock contents
-    internalDockWidget = QWidget(dockWidgetContents)
-
-    # Add vertical layout to dock widget
-    internalWidgetVerticalLayout = QVBoxLayout(internalDockWidget)
-    internalWidgetVerticalLayout.setContentsMargins(0, 0, 0, 0)
-
-    # Add group box
-    paramsGroupBox = QGroupBox(internalDockWidget)
-
-
-    # Add form layout to group box
-    groupBoxFormLayout = QFormLayout(paramsGroupBox)
-    #groupBoxFormLayout.setFormAlignment(Qt.AlignCenter)
-
-    # Add elements to layout
-    internalWidgetVerticalLayout.addWidget(paramsGroupBox)
-    dockContentsVerticalLayout.addWidget(internalDockWidget)
-    dockWidget.setWidget(dockWidgetContents)
-
-    #        self.graphWidgetVL.addWidget(self.graphParamsGroupBox)
-    #        self.graphDockVL.addWidget(self.dockWidget)
-    #        self.pointCloudDockWidget.setWidget(self.pointCloudDockWidgetContents)
-    #
-    # self.addDockWidget(QtCore.Qt.RightDockWidgetArea, self.pointCloudDockWidget)
-    return (dockWidget, dockWidgetContents,
-            dockContentsVerticalLayout, internalDockWidget,
-            internalWidgetVerticalLayout, paramsGroupBox,
-            groupBoxFormLayout)
-
-
-def main():
-    app = QtWidgets.QApplication([])
-    
-    file_dir = os.path.dirname(__file__)
-    owl_file = os.path.join(file_dir, "DVCIconSquare.png")
-    owl = QtGui.QPixmap(owl_file)
-    splash = QtWidgets.QSplashScreen(owl)
-    splash.show()
-    
-    err = vtk.vtkFileOutputWindow()
-    err.SetFileName("../viewer.log")
-    vtk.vtkOutputWindow.SetInstance(err)
-
-    
-    window = MainWindow()
-    
-    window.show()
-    splash.finish(window)
-
-    sys.exit(app.exec_())
-
-if __name__ == "__main__":
-    main()
